Using Consumer
Consumer widget we use if we dont want to build whole widget tree anf only widget which is changed.
for that in provider we do (listen:false) and wrap widget which is getting changed with consumer.
it also has child properly in which we can define a child of that widget which doesnts changes and
we dont want rebuild it again.